// src/api/video-v1/model/subtitles.ts
/**
 * Drtv Videos Service API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Subtitles { 
    language: string;
    format: string;
    isEmbedded: boolean;
    isHardSubs: boolean;
    link: string | null;
}